interactive-flask-app/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv==0.19.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
requests==2.28.1
pytest==7.1.2
bcrypt==3.2.0